+++ business/modules/nlp_test.go
@@ -1,4 +1,4 @@
-package business
+package modules
 
 import (
 	"testing"
